Fixing errors when using pre compiled binaries
