+++ action/hit_ground_ball/hit_ground_ball.py
@@ -44,20 +44,21 @@ class HitGroundBall(BaseAction):
         origin_height = 17  # the car's elevation from the ground due to wheels and suspension
 
         # only accurate if we're already moving towards the target
-        boost = np.array([car.boost] * len(ball_prediction))
+        boost = np.array([car.boost] * len(ball_prediction), dtype=np.float64)
 
         location_slices = ball_prediction["physics"]["location"]
 
         distance_slices = box_ball_collision_distance(
             location_slices, car.location, car_rot, car.hitbox_corner, car.hitbox_offset, ball.radius,
         )
-        time_slices = ball_prediction["game_seconds"] - game_data.time
+        time_slices = np.array(ball_prediction["game_seconds"] - game_data.time, dtype=np.float64)
 
         not_too_high = location_slices[:, 2] < ball.radius + hitbox_height + origin_height
 
         velocity = car.velocity[None, :]
         direction_slices = location_slices - car.location
         velocity = np.sum(velocity * direction_slices, 1) / np.linalg.norm(direction_slices, 2, 1)
+        velocity = np.array(velocity, dtype=np.float64)
 
         reachable = (state_at_time_vectorized(time_slices, velocity, boost)[0] > distance_slices) & not_too_high
